# Conflicts:
#	linkwe-service/src/main/java/com/linkwechat/domain/WeCorpAccount.java
#	linkwe-service/src/main/java/com/linkwechat/fallback/QwCustomerFallbackFactory.java
#	linkwe-service/src/main/java/com/linkwechat/fegin/QwCustomerClient.java
#	linkwe-wecom/src/main/java/com/linkwechat/wecom/controller/QwCustomerController.java
#	linkwe-wecom/src/main/java/com/linkwechat/wecom/service/IQwCustomerService.java
#	linkwe-wecom/src/main/java/com/linkwechat/wecom/service/impl/QwCustomerServiceImpl.java

+++ linkwe-service/src/main/java/com/linkwechat/domain/qr/vo/WeLxQrScopeUserVo.java
@@ -23,7 +23,7 @@ public class WeLxQrScopeUserVo {
     private String userName;
 
     @ApiModelProperty(value = "部门id")
-    private Integer party;
+    private Long party;
 
     @ApiModelProperty(value = "部门名称")
     private String partyName;
